+++ NEWS.md
@@ -21,6 +21,7 @@
   upgrades it with `upgradeParams()` if necessary.
 * Upgrade old params objects automatically when used in plot functions, rate
   functions, summary functions or in `project()` or `steady()`, #163.
+* New function `getRates()` to calculate all rates and collect them in a list.
 * Numerous small improvements to documentation.
 
 # mizer 2.0.3